// packages/unlighthouse-old-mono/src/types.ts
export interface RouteDefinition {
  name: string
  path: string
  file?: string
  component?: string
  componentBaseName?: string
  chunkName?: string
  _name?: string
  layout?: string
}

/**
 * A normalised route, in the context of unlighthouse, is a mapping between a URL path, it's definition and a unique id.
 */
export interface NormalisedRoute {
  id: string
  path: string
  url: string
  $url: URL
  definition: RouteDefinition
  /**
   * A runtime path that the route was discovered from, useful if the route is a 404 and we want to know what directed
   * us to it.
   */
  discoveredFrom?: string
}

export interface ComputedLighthouseReportAudit {
  details?: {
    items?: any[]
  }
  displayValue: string | number
  score: number
}
/**
 * An augmented Lighthouse Report type, we add custom types to the base report for specific functionality on the
 * @unlighthouse/client.
 */
export type LighthouseReport = Partial<Result> & {
  /**
   * The total score for the result, this is the sum of each category's result
   */
  score: number
  categories: { score: number | null }[]
  computed: {
    /**
     * An aggregation of multiple image audit results.
     */
    imageIssues: ComputedLighthouseReportAudit
    ariaIssues: ComputedLighthouseReportAudit
  }
}

/**
 * Tasks that Unlighthouse will run, used to track their status.
 */
export type UnlighthouseTask = 'inspectHtmlTask' | 'runLighthouseTask'

/**
 * Each task ran by unlighthouse (extractHtmlPayload, runLighthouseTask) has a specific status which we can expose.
 */
export type UnlighthouseTaskStatus = 'waiting' | 'in-progress' | 'completed' | 'failed' | 'ignore' | 'failed-retry'

/**
 * A fairly rigid representation of the puppeteer cluster task results (extractHtmlPayload, runLighthouseTask), combined
 * with the normalised route.
 */
export interface UnlighthouseRouteReport {
  /**
   * The mapping of tasks with their status.
   */
  tasks: Record<UnlighthouseTask, UnlighthouseTaskStatus>
  /**
   * Track how long tasks are taking
   */
  tasksTime?: Record<Partial<UnlighthouseTask>, number>
  /**
   * Path to where the artifacts from a URL scan are saved.
   */
  artifactPath: string
  /**
   * URL of where the artifacts are stored, for static client access.
   */
  artifactUrl: string
  /**
   * The route (URL Path) that the report belongs to.
   */
  route: NormalisedRoute
  /**
   * A unique representation of the route, useful for the API layer.
   */
  reportId: string
  /**
   * The lighthouse result, only set once the task is completed.
   */
  report?: LighthouseReport
  /**
   * The SEO meta-data, only set once the html payload has been extracted and passed.
   */
  seo?: HTMLExtractPayload
}

export interface HTMLExtractPayload {
  alternativeLangDefault?: string
  alternativeLangDefaultHtml?: string
  title?: string
  description?: string
  internalLinks?: number
  externalLinks?: number
  htmlSize?: number
  favicon?: string
  og?: {
    description?: string
    title?: string
    image?: string
  }
}
// ...
export interface MockRouter { match: (path: string) => RouteDefinition | false }
// ...
export interface GenerateClientOptions {
  static?: boolean
}
// ...
export type ClientOptionsPayload = Pick<ResolvedUserConfig, 'client' | 'site' | 'lighthouseOptions' | 'scanner' | 'routerPrefix'>
  & Pick<RuntimeSettings, 'websocketUrl' | 'apiUrl'>

type DeepPartial<T> = {
  [K in keyof T]?: T[K] extends object ? DeepPartial<T[K]> : T[K]
}

export type UserConfig = DeepPartial<ResolvedUserConfig>

export interface RuntimeSettings {
  /**
   * A URL instance of the site for easier use of the host.
   */
  siteUrl: URL
  /**
   * The URL of the server running the API and client.
   */
  serverUrl: string
  /**
   * The full URL for API.
   */
  apiUrl: string
  /**
   * The path of the api without the site URL.
   */
  apiPath: string
  /**
   * Whether we have managed to resolve definitions for the routes.
   */
  hasRouteDefinitions: boolean
  /**
   * if the user has configured unlighthouse using a configuration file, this is the link to it.
   */
  configFile?: string
  /**
   * When caching we need to generate a unique key based on config.
   */
  configCacheKey?: string
  /**
   * The URL that the websocket will be served from, depends on the api.prefix. Will look something like:
   * "ws://localhost:3000/ws".
   */
  websocketUrl: string
  /**
   * The resolved local path to the generated client.
   */
  generatedClientPath: string
  /**
   * The URL to the client, used for opening it automatically.
   */
  clientUrl: string
  /**
   * The resolved local path to the client dist.
   */
  resolvedClientPath: string
  /**
   * The resolved output path we'll be saving reports and the client in.
   */
  outputPath: string
  /**
   * The root directory of the module.
   */
  moduleWorkingDir: string
  /**
   * The path to the lighthouse worker.
   */
  lighthouseProcessPath: string

  /**
   * The server instance.
   */
  server: http.Server | https.Server
}

export interface UnlighthouseWorkerStats {
  /**
   * Status of the worker, completed when all tasks have been completed.
   */
  status: 'completed' | 'working'
  /**
   * Time in ms that the worker has been running
   */
  timeRunning: number
  /**
   * How many tasks have been completed.
   */
  doneTargets: number
  /**
   * Total number of tasks including completed, pending and working.
   */
  allTargets: number
  /**
   * The % of work completed.
   */
  donePercStr: string
  /**
   * The % of errors.
   */
  errorPerc: string
  /**
   * The remaining time until all tasks are completed.
   */
  timeRemaining: number
  /**
   * How many tasks per second are being processed.
   */
  pagesPerSecond: string
  /**
   * The devices CPU usage % out of 100
   */
  cpuUsage: string
  /**
   * The devices memory usage % out of 100
   */
  memoryUsage: string
  /**
   * How many workers are now working, usually the cpu count of the device.
   */
  workers: number
}
// ...
export interface Provider {
  /**
   * Used to debug.
   */
  name?: string
  /**
   * To match a URL path to a route definition we need a router. Different definitions need different routes.
   */
  mockRouter?: MockRouter | ((routeDefinitions: RouteDefinition[]) => MockRouter)
  /**
   * The collection of route definitions belonging to the provider. These can be inferred but aren't 100% correct,
   * frameworks that can provide these should do so.
   */
  routeDefinitions?: RouteDefinition[] | (() => RouteDefinition[] | Promise<RouteDefinition[]>)
}

export type HookResult = Promise<void> | void

export interface UnlighthouseHooks {
  /**
   * It's possible the site is not known at initialisation, this hook is called when it's set or changed.
   * @param site The site that was set.
   */
  'site-changed': (site: string) => HookResult
  /**
   * Once the config is resolved.
   *
   * @param resolvedConfig
   */
  'resolved-config': (resolvedConfig: ResolvedUserConfig) => HookResult
  /**
   * Called when the worker has finished processing all queued routes. Will be called multiple times if routes are
   * re-queued.
   *
   * Mostly useful for the CI environment.
   */
  'worker-finished': () => HookResult
  /**
   * When route definitions are provided to Unlighthouse this function will be called, useful for delaying internal logic
   * until the definitions are found.
   *
   * @param routeDefinitions
   */
  'route-definitions-provided': (routeDefinitions: any[]) => HookResult
  /**
   * Called when a user visits the path of the @unlighthouse/client for the first time. Useful for starting the worker on-demand.
   */
  'visited-client': () => HookResult
  /**
   * Fired when a new task is added to the queue worker.
   * @param path
   * @param response
   */
  'task-added': (path: string, response: UnlighthouseRouteReport) => HookResult
  /**
   * Fired when a task has started to work.
   * @param path
   * @param response
   */
  'task-started': (path: string, response: UnlighthouseRouteReport) => HookResult
  /**
   * Fired when a task has completed it's work.
   * @param path
   * @param response
   */
  'task-complete': (path: string, response: UnlighthouseRouteReport, taskName: string) => HookResult
  /**
   * Fired when a path discovered internal links, used for "crawl" mode.
   * @param path
   * @param internalLinks
   */
  'discovered-internal-links': (path: string, internalLinks: string[]) => HookResult
  /**
   * After a page has been visited with puppeteer. Useful for running
   * @param page
   */
  'puppeteer:before-goto': (page: Page) => HookResult
  /**
   * Authenticate a page before it's visited.
   * @param page
   */
  'authenticate': (page: Page) => HookResult
}

/**
 * The worker is the manager of puppeteer-cluster, it provides an API for managing the queued routes, monitoring the queue and
 * fetching the results of the queued jobs.
 */
export interface UnlighthouseWorker {
  /**
   * puppeteer-cluster instance
   */
  cluster: Cluster
  /**
   * A collection of stats gathered from the cluster for the current status of the worker.
   */
  monitor: () => UnlighthouseWorkerStats
  /**
   * Queue a single normalised route. Will not process routes that have already been queued.
   * @param route
   */
  queueRoute: (route: NormalisedRoute) => void
  /**
   * Queue multiple normalised routes. This will sort the list for a better loading experience.
   * @param routes
   */
  queueRoutes: (routes: NormalisedRoute[]) => void
  /**
   * Re-queues a report, avoiding the usual caching involved and makes sure we unlink any of the previous reports data or
   * tasks.
   *
   * @param report
   */
  requeueReport: (report: UnlighthouseRouteReport) => void
  /**
   * Has the worker started processing the queue.
   */
  hasStarted: () => boolean
  /**
   * The gathered map of reports. The key is the path of the route.
   */
  routeReports: Map<string, UnlighthouseRouteReport>
  /**
   * A simple array representation of the reports for easy iteration.
   */
  reports: () => UnlighthouseRouteReport[]
  /**
   * Find a report with the specified id.
   * @param id
   */
  findReport: (id: string) => UnlighthouseRouteReport | null

  /**
   * Iterates through route reports checking for a match on the route definition component, if there is a match
   * then the route is re-queued.
   *
   * @param file
   * @return True if an invalidation occurred on the routes.
   */
  invalidateFile: (file: string) => boolean

  /**
   * Checks if the amount of routes scanned surpasses the `scanner.maxRoutes` value.
   */
  exceededMaxRoutes: () => boolean

  /**
   * Clear the progress display box
   */
  clearProgressDisplay: () => void
}

export interface ScanMeta {
  /**
   * Total count of discovered routes
   */
  routes: number
  /**
   * How are worker is operating
   */
  monitor?: UnlighthouseWorkerStats
  /**
   * Aggregate score for the site
   */
  score: number
  /**
   * Discovered site favicon.
   */
  favicon?: string
}

export interface ServerContextArg {
  url: string
  server: http.Server | https.Server
  app: any /* can't type h3 here */
}

/**
 * The context is provided by the createUnlighthouse() or useUnlighthouse() functions.
 * It provides the central API to interact with the behaviour of Unlighthouse..
 */
export interface UnlighthouseContext {
  /**
   * The mock router being used to match paths to route definitions.
   */
  mockRouter?: MockRouter
  /**
   * Settings that are computed from runtime data.
   */
  runtimeSettings: RuntimeSettings
  /**
   * Access the hook system, either calling a hook or listening to one.
   */
  hooks: Hookable<UnlighthouseHooks>
  /**
   * User config that has been normalised.
   */
  resolvedConfig: ResolvedUserConfig
  /**
   * The collection of route definitions associated to the site.
   */
  routeDefinitions?: RouteDefinition[]
  /**
   * Discovered routes.
   */
  routes?: NormalisedRoute[]
  /**
   * A reference to the API middleware.
   */
  api: any
  /**
   * A reference to the websocket interface, used to broadcast data.
   */
  ws: WS
  /**
   * Access the worker environment, queue tasks, inspect progress, etc.
   */
  worker: UnlighthouseWorker
  /**
   * Provider details
   */
  provider: Provider

  /**
   * To use Unlighthouse with a client, it needs a server / app to register the API and client middleware.
   *
   * @param arg
   */
  setServerContext: (arg: ServerContextArg) => Promise<UnlighthouseContext>
  /**
   * Sets the site URL that will be scanned if it's not known at initialisation.
   * @param url
   */
  setSiteUrl: (url: string) => void
  /**
   * Running Unlighthouse via CI does not require a server or the client so we have a special utility for it.
   */
  setCiContext: () => Promise<UnlighthouseContext>
  /**
   * Start the client and the queue worker. A server context must be provided before this function is called.
   */
  start: () => Promise<UnlighthouseContext>
  /**
   * @internal
   */
  _axios?: AxiosInstance
  /**
   * @internal
   */
  _i18nWarn?: boolean
}